+++ piwigoPublish.lrplugin/PublishDialogSections.lua
@@ -238,11 +238,11 @@ local function prefsDialog (f, propertyTable)
 			f:push_button {
 				title = 'Create Special Collections',
 				width = share 'buttonwidth',
-				--enabled = bind('Connected', propertyTable),
-				enabled = false, -- temporary disabled
+				enabled = bind('Connected', propertyTable),
+				--enabled = false, -- temporary disabled
 				tooltip = "Create special publish collections for publish collection sets, allowing images to be published to Piwigo albums with sub-albums",
 				action = function(button)
-					local result = LrDialogs.confirm("Create Special Collections","Are you sure you want to create Special Collections?\nExisting collections will be unaffected.","Import","Cancel")
+					local result = LrDialogs.confirm("Create Special Collections","Are you sure you want to create Special Collections?\nExisting collections will be unaffected.","Create","Cancel")
 					if result == 'ok' then
 						LrTasks.startAsyncTask(function()
 							PiwigoAPI.specialCollections(propertyTable)